feat: yaml interface
