+++ main.py
@@ -1,9 +1,11 @@
-from surfaceid.util.utils import (SEARCH, ALIGN, DESC, RLIM, SIG,  Aligner1, get_fs,
+from surfaceid.util.utils import (RLIM, SIG, get_fs,
                                   get_whole_molecule_desc, get_fs_npz, get_xyz, get_centroid,
                                   get_score1, get_score2, get_normal_npz, get_xyz_npz, transform_library,
                                   Model, compute_aux_vars, gen_descriptors_contact, get_desc_aux,
                                   search, get_within)
 
+from surfaceid.util.Config import Config
+from surfaceid.model.aligner import Aligner1
 
 import logging
 import os
@@ -24,166 +26,135 @@ from plyfile import PlyData, PlyElement
 from copy import deepcopy
 import shutil
 
-torch.set_num_threads(cpu_count())
-CONTACT = False
-DESC = False
-SEARCH = False
-ALIGN = False
-SAVEPLY = False
-HEATMAP = False
-######################################
-# Parameter
-######################################
-REMOVE_OLD = ALIGN and SAVEPLY and SEARCH
-CATALOG_PAIRS = "data/20201107_SAbDab_catalog_pairs.tsv"
-
-expand_radius = 2.0
-neighbor_dist = 3.0
-nmin_pts = 40  # for grouping target decoys
-nmin_pts_library = 30  # library epitope
-# Model
-prefix = "final_002"
-device = 'cpu'
-
-# ---- Identify antigens
-case = "4fqi_AB"
-OUTDIR_RESULTS = f"{case}_results"
-OUTDIR = "data/20201107_SAbDab_masif"
-target = "4fqi_AB"
-
-contact_thres1 = 3.0  # for identifying iface points
-contact_thres2 = None  # for distance based interacting points
-contact_mode = "dist"
-
-npatience = 100
-
-# --- 6AA
-params = {"rho_max": 6.0, "nbins_rho": 5, "nbins_theta": 16, "num_in": 5,
-          "neg_margin": 10, "add_center_pixel": True,
-          "share_soft_grid_across_channel": True,
-          "conv_type": "sep", "num_filters": 512, "weight_decay": 1e-2,
-          "dropout": 0.1, "min_sig": 5e-2, "lr": 5e-4}
-thres = 4.0
-
-savedir = "models"
 
 logging.basicConfig(stream=sys.stdout, level=logging.INFO)
 logger = logging.getLogger("🧪 SurfaceID")
 
-parser = argparse.ArgumentParser(
+argparser = argparse.ArgumentParser(
     prog="python main.py ...",
     formatter_class=argparse.RawDescriptionHelpFormatter,
     description=(
         """\
-                
+
                 -----------------------------------------------------
                 Surface Similarity Search via Geometric Deep Learning
                 -----------------------------------------------------
         """
     ),
 )
-parser.add_argument(
-    "--pdb", default="4fqi_AB", help="Input file for execution"
-)
-parser.add_argument(
+argparser.add_argument(
     "--device", default="cpu", help="Device to compute"
 )
-parser.add_argument(
+argparser.add_argument(
     "--params", default="config/config_4fqi.yml", help="config file with model parameter"
 )
 # parse the arguments
-(args) = parser.parse_args()
+(args) = argparser.parse_args()
 
 config = yaml.safe_load(open(args.params, "r"))
-logger.info(config)
+logger.info(f"Parsing config file {args.params}")
 
-if torch.cuda.is_available():
+conf=Config(args.params)
+
+if args.device == 'cpu':
+    torch.set_num_threads(cpu_count())
+
+elif torch.cuda.is_available():
     device = torch.cuda.current_device()
 
-if SEARCH or DESC or ALIGN:
-    path = os.path.join(savedir, f"model_{prefix}.pth")
-    model = Model(**params)
+if conf.SEARCH or conf.DESC or conf.ALIGN:
+    path = os.path.join(conf.savedir, f"model_{conf.prefix}.pth")
+    model = Model(**conf.params)
     model = model.to(device)
     model.dthetas = model.dthetas.to(device)
     model.load_state_dict(torch.load(
         path, map_location=torch.device('cpu')))
     model.eval()
 
-df = pd.read_csv(CATALOG_PAIRS, sep="\t")
+df = pd.read_csv(conf.CATALOG_PAIRS, sep="\t")
 df.Ab = df.apply(lambda x: f'{x.id}_{x.Ab}', axis=1)
 df.Ag = df.apply(lambda x: f'{x.id}_{x.Ag}', axis=1)
 logger.info(f"# of pairs: {len(df)}")
 chains = set(df.Ab.tolist() + df.Ag.tolist())
 logger.info(f"# of chains: {len(chains)}")
 chains_processed = [chain for chain in chains if os.path.exists(
-    f"{OUTDIR}/{chain}_surface.npz")]
+    f"{conf.OUTDIR}/{chain}_surface.npz")]
 logger.info(
     f"# of chains processed: {len(chains_processed)}")
 df = df[df.apply(lambda x: (x["Ag"] in chains_processed) and (
     x.Ab in chains_processed), axis=1)].reset_index(drop=True)
 logger.info(f"# of pairs after filter: {len(df)}")
 
-if os.path.exists(OUTDIR_RESULTS) and REMOVE_OLD:
-    shutil.rmtree(OUTDIR_RESULTS)
-os.makedirs(OUTDIR_RESULTS, exist_ok=True)
+if os.path.exists(conf.OUTDIR_RESULTS) and conf.REMOVE_OLD:
+    shutil.rmtree(conf.OUTDIR_RESULTS)
+os.makedirs(conf.OUTDIR_RESULTS, exist_ok=True)
 
-if CONTACT:
+if conf.CONTACT:
     for i, x in df.iterrows():
         p1 = x.Ab
         p2 = x.Ag
         logger.info(f"{i}, {p1}, {p2}")
-        compute_aux_vars(p1, p2, OUTDIR, expand_radius, neighbor_dist,
-                         contact_mode, contact_thres1, contact_thres2, device, smooth=False)
+        compute_aux_vars(p1,
+                         p2, 
+                         conf.OUTDIR, 
+                         conf.expand_radius, 
+                         conf.neighbor_dist,
+                         conf.contact_mode, 
+                         conf.contact_thres1, 
+                         conf.contact_thres2,
+                         device, 
+                         smooth=False)
 
-if DESC:
+if conf.DESC:
     p1s = df.Ab.to_numpy()
     p2s = df.Ag.to_numpy()
-    fname = f"nn_desc.{case}.npz"
+    fname = f"nn_desc.{conf.case}.npz"
     pdbs, x_desc, idxs_desc = gen_descriptors_contact(
-        p1s, p2s, model, device, params["rho_max"], OUTDIR, fname)
+        p1s, p2s, model, device, conf.params.get("rho_max"), conf.OUTDIR, fname)
     patch_sizes, pdbs_unique, starts, ends = get_desc_aux(pdbs)
 
 # ---------------- All vs all search
 # Load descriptors
 # For each Ag epitope, compare against all other
-if not DESC and (ALIGN or SEARCH):
-    fname = f"nn_desc.{case}.npz"
+if not conf.DESC and (conf.ALIGN or conf.SEARCH):
+    fname = f"nn_desc.{conf.case}.npz"
     data = np.load(fname)
     pdbs = data.pdbs
     x_desc = data.x
     idxs_desc = data["idxs"]
     patch_sizes, pdbs_unique, starts, ends = get_desc_aux(pdbs)
 
-if SEARCH:
+if conf.SEARCH:
 
     # libraries
     library = df.apply(lambda x: f"{x['Ag']}.{x['Ab']}", axis=1).to_list()
-    library_within = [np.load(os.path.join(OUTDIR, x)) for x in df.apply(
+    library_within = [np.load(os.path.join(conf.OUTDIR, x)) for x in df.apply(
         lambda x: f"{x['Ag']}_within.{x['Ab']}.npy", axis=1).to_list()]
-    DIM = params["num_filters"]
+    DIM = conf.params.get("num_filters")
     n = len(pdbs_unique)
     summary = []
-    target = "4fqi_AB"
     st_target = time()
     o1, p1 = get_whole_molecule_desc(
-        target, model, params["rho_max"], bs=128, device=device, outdir=OUTDIR)
+        target, model, conf.params.get("rho_max"), bs=128, device=device, outdir=conf.OUTDIR)
     o1 = torch.tensor(o1).to(device).view(len(o1), 1, DIM)
     mini_bs = int(64 * 1000 / len(o1))
     rho = p1.rho.cpu().numpy()
     list_indices = p1.list_indices.cpu().numpy()
 
     # Get inclusion matrix
-    within = get_within(rho, list_indices, neighbor_dist)
+    within = get_within(rho, 
+                        list_indices, 
+                        conf.neighbor_dist)
+    
     summary.extend(search(o1, rho, list_indices, within, library, library_within, pdbs_unique, patch_sizes,
-                          pdbs, x_desc, idxs_desc, mini_bs, device, nmin_pts, expand_radius, nmin_pts_library,
+                          pdbs, x_desc, idxs_desc, mini_bs, device, conf.nmin_pts, conf.expand_radius, conf.nmin_pts_library,
                           thres, target, idxs_contact=None))
     # remove o1
     o1 = o1.cpu()
     del o1
-    (f"{target} / dt = {time() - st_target:.3f}")
+    (f"{conf.target} / dt = {time() - st_target:.3f}")
 
-    df_hits = pd.DataFrame(summary, columns=["target", "library epitope", "target_nhits", "target_nexpanded", "mean_desc_dist",
-                                             "library_nhits", "library_nexpanded", "frac_library_hits", "target_expanded_indices", "library_epitope_indices"])
+    df_hits = pd.DataFrame(summary, columns=conf.HIT_COLUMNS)
     df_hits["frac_expanded"] = df_hits["target_nhits"] / \
         df_hits["target_nexpanded"]
     df_hits["frac_geo"] = np.sqrt(
@@ -193,28 +164,27 @@ if SEARCH:
     df_hits["library_epitope_indices"] = df_hits["library_epitope_indices"].apply(
         lambda x: list(x))
     df_hits.to_csv(os.path.join(
-        OUTDIR_RESULTS, f"{case}_top_hits.tsv"), sep="\t", index=False)
+        conf.OUTDIR_RESULTS, f"{conf.case}_top_hits.tsv"), sep="\t", index=False)
 
 if ALIGN:
     df_hits = pd.read_csv(os.path.join(
-        OUTDIR_RESULTS, f"{case}_top_hits.tsv"), sep="\t")
+        conf.OUTDIR_RESULTS, f"{conf.case}_top_hits.tsv"), sep="\t")
     df_hits["target_expanded_indices"] = df_hits["target_expanded_indices"].apply(
         lambda x: np.asarray(eval(x), dtype=np.int32))
     df_hits["library_epitope_indices"] = df_hits["library_epitope_indices"].apply(
         lambda x: np.asarray(eval(x), dtype=np.int32))
     df_hits["id"] = -1
-    npatience = 100
+    conf.npatience = 100
     DIM = params["num_filters"]
     librarys = df.apply(lambda x: f'{x["Ag"]}.{x["Ab"]}', axis=1).tolist()
-    target = "4fqi_AB"
     id_counter = 0
     st_target = time()
-    logger.info(f"/----{target}")
-    if SAVEPLY:
+    logger.info(f"/----{conf.target}")
+    if conf.SAVEPLY:
         ply_target = deepcopy(PlyData.read(
-            os.path.join(OUTDIR, f"{target}.ply")))
+            os.path.join(conf.OUTDIR, f"{conf.target}.ply")))
     o_target, p_target = get_whole_molecule_desc(
-        target, model, params["rho_max"], bs=128, device=device, outdir=OUTDIR)
+        target, model, params["rho_max"], bs=128, device=device, outdir=conf.OUTDIR)
     mini_bs = int(64 * 1000 / len(o_target))
     for library in librarys:
         df_ = df_hits[(df_hits["target"] == target) & (
@@ -225,9 +195,7 @@ if ALIGN:
         for ie, entry in df_.iterrows():
             st = time()
             library_p1, library_p2 = entry.library.epitope.split(".")
-            # assert False
-            # Load library data
-            fname = os.path.join(OUTDIR, f"{library_p1}_surface.npz")
+            fname = os.path.join(conf.OUTDIR, f"{library_p1}_surface.npz")
             fs, idxs_library0 = get_fs(
                 entry["library epitope"], pdbs, x_desc, idxs_desc)
             # Get a subselection of match points
@@ -241,10 +209,10 @@ if ALIGN:
             xyz_mean = get_centroid(xyz)
             xyz -= xyz_mean
 
-            fname_target = os.path.join(OUTDIR, f"{target}_surface.npz")
+            fname_target = os.path.join(conf.OUTDIR, f"{conf.target}_surface.npz")
             idxs_target = entry["target_expanded_indices"]
 
-            if (len(idxs_target) < nmin_pts) or (len(idxs_target) > 5000):
+            if (len(idxs_target) < conf.nmin_pts) or (len(idxs_target) > 5000):
                 continue
             fs0 = o_target[idxs_target]
             fs_raw0 = get_fs_npz(fname_target, idxs_target)
@@ -266,7 +234,7 @@ if ALIGN:
                     d = torch.norm(fs0 - fs_, dim=2)
                     ds.append(d)
                 d = torch.cat(ds, dim=1)
-                mask = (d < thres).float()
+                mask = (d < conf.thres).float()
             T_zero = np.asarray([0, 0, 0])
 
             # -- 1st alignment
@@ -300,7 +268,7 @@ if ALIGN:
                     min_loss = loss
                 else:
                     nbad += 1
-                if (nbad == npatience):
+                if (nbad == conf.npatience):
                     print(f"1 / {i} loss_min", losses.min().item(), idx1)
                     break
                 if i == 0:
@@ -330,16 +298,14 @@ if ALIGN:
             df_hits.loc[ie, "id"] = id_counter
 
             logger.info(f"Time for {library}: {time()-st:.2f}")
-
-            # and ((df_hits.loc[ie, f"aln1_score2"] < 0.9) or (df_hits.loc[ie, f"frac_library_hits"] > 0.3)):
-            if SAVEPLY:
-                fname_library = os.path.join(OUTDIR, f"{library_p1}.ply")
+            if conf.SAVEPLY:
+                fname_library = os.path.join(conf.OUTDIR, f"{library_p1}.ply")
                 if not os.path.exists(fname_library):
                     continue
 
                 # -- create ply after aln1
                 # transform contact patch
-                fname = os.path.join(OUTDIR, f"{library_p1}_surface.npz")
+                fname = os.path.join(conf.OUTDIR, f"{library_p1}_surface.npz")
                 els = []
                 for tag, iis in zip(["hit_patch", "contact_patch"], [idxs_library, idxs_library0]):
                     xyzp = get_xyz_npz(fname, iis)
@@ -357,7 +323,7 @@ if ALIGN:
                     xp, yp, zp, xyz0_mean, xyz_mean, aligner1, idx1, n=None)
                 # Alter vertices and save
                 fname = os.path.join(
-                    OUTDIR_RESULTS, f"{id_counter}_{entry['library epitope']}_{target}.ply")
+                    conf.OUTDIR_RESULTS, f"{id_counter}_{entry['library epitope']}_{target}.ply")
                 ply_library["vertex"]["x"] = xyzp[:, 0]
                 ply_library["vertex"]["y"] = xyzp[:, 1]
                 ply_library["vertex"]["z"] = xyzp[:, 2]
@@ -375,7 +341,7 @@ if ALIGN:
 
                 # ---- Alter and save pdb
                 # Parse the structure file
-                fname = os.path.join(OUTDIR, f"{library_p1}.pdb")
+                fname = os.path.join(conf.OUTDIR, f"{library_p1}.pdb")
                 if not os.path.exists(fname):
                     continue
                 parser = PDBParser()
@@ -389,17 +355,17 @@ if ALIGN:
                         xp, yp, zp, xyz0_mean, xyz_mean, aligner1, idx1, n=None)
                     atom.set_coord(xyzp[0])
                 fname = os.path.join(
-                    OUTDIR_RESULTS, f"{id_counter}_{entry['library epitope']}_{target}.pdb")
+                    conf.OUTDIR_RESULTS, f"{id_counter}_{entry['library epitope']}_{target}.pdb")
                 io = PDBIO()
                 io.set_structure(structure)
                 io.save(fname)
-    logger.info(f"Time for {target}: {time()-st_target:.2f}")
+    logger.info(f"Time for {conf.target}: {time()-st_target:.2f}")
     if SAVEPLY:
         out_target_fname = os.path.join(
-            OUTDIR_RESULTS, f"{target}_ref.ply")
+            conf.OUTDIR_RESULTS, f"{conf.target}_ref.ply")
         ply_target.write(out_target_fname)
     df_hits.to_csv(os.path.join(
-        OUTDIR_RESULTS, f"{case}_top_hits_aligned.tsv"), sep="\t", index=False)
+        conf.OUTDIR_RESULTS, f"{conf.case}_top_hits_aligned.tsv"), sep="\t", index=False)
 
     # ---- Ranking vs. scores
     df_ = df_hits[(~df_hits["aln1_score2"].isna())]
@@ -414,7 +380,7 @@ if ALIGN:
             f"mean feat dist of closest vertex pairs ({RLIM}) cutoff")
         ax.legend(loc="upper right")
         plt.savefig(os.path.join(
-            OUTDIR_RESULTS, f"{case}_{rank}_vs_score2.png"), dpi=200, bbox_inches="tight")
+            conf.OUTDIR_RESULTS, f"{conf.case}_{rank}_vs_score2.png"), dpi=200, bbox_inches="tight")
         plt.close()
         ascending = True if rank == "mean_desc_dist" else False
         df__ = df_.sort_values(
@@ -423,18 +389,17 @@ if ALIGN:
         nhit = np.sum(iselect)
         logger.info(rank, nhit)
 
-    # change var name
     df_hits = pd.read_csv(os.path.join(
-        OUTDIR_RESULTS, f"{case}_top_hits_aligned.tsv"), sep="\t")
+        conf.OUTDIR_RESULTS, f"{conf.case}_top_hits_aligned.tsv"), sep="\t")
     cols = ["library epitope", "aln1_score2",
             "aln1_score3", "aln1_ncontact", "aln1_normal_score"]
     cols_new = ["library_epitope", "spatial_sim",
                 "spatial_sim_desc", "ncontact", "normal_score"]
     df_hits.rename(
         columns={c: cn for c, cn in zip(cols, cols_new)}, inplace=True)
-    fname = os.path.join(OUTDIR_RESULTS, f"{case}_top_hits_aligned.tsv")
+    fname = os.path.join(conf.OUTDIR_RESULTS, f"{conf.case}_top_hits_aligned.tsv")
     df_hits.to_csv(fname, sep="\t", index=False)
-    df_annots = pd.read_csv(CATALOG_PAIRS, sep="\t")
+    df_annots = pd.read_csv(conf.CATALOG_PAIRS, sep="\t")
     df_annots1 = deepcopy(df_annots)
     df_annots1["id"] = df_annots1["id"] + "_" + df_annots1["Ag"] + \
         "." + df_annots1["id"] + "_" + df_annots1["Ab"]
@@ -462,6 +427,6 @@ if ALIGN:
             iselect = (df__["spatial_sim"] < sim_thres) & (
                 df__["ncontact"] > ncontact_thres) & ~ihem
             nhit_nonha = np.sum(iselect)
-            logger.info(f"Top {mm}", nhit_ha, nhit_nonha)
+            logger.info(f"Top {mm}, {nhit_ha}, {nhit_nonha}")
             logger.info((df__[iselect][cols_new + [rank, "compound",
                         "antigen_name", "id"]].sort_values("ncontact")))
